requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
google-genai
jinja2
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
mutagen
python-magic
